+++ packages/carousel/src/react/context.js
@@ -1,6 +1,9 @@
 import React from 'react'
 
-const initialValue = {}
+const initialValue = {
+  transitioning: false,
+  setTransitioning: () => {}
+}
 const context = React.createContext(initialValue)
 
 export default context