+++ packages/navitem/.storybook/addons.js
@@ -1,0 +1,1 @@
+import '@pluralsight/ps-design-system-storybook-addon-theme/register'