+++ packages/navitem/src/vars/index.js
@@ -1,0 +1,4 @@
+export const alignments = {
+  horizontal: 'horizontal',
+  vertical: 'vertical'
+}