+++ packages/site/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [19.0.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.3...@pluralsight/ps-design-system-site@19.0.4) (2020-07-03)
+
+
+### Bug Fixes
+
+* **table:** doubled up click handler nulling sort ([#1049](https://github.com/pluralsight/design-system/issues/1049)) ([2b02663](https://github.com/pluralsight/design-system/commit/2b0266309a7a4b673940eed21a1c4b74e960ca11))
+
+
+
+
+
+## [19.0.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.2...@pluralsight/ps-design-system-site@19.0.3) (2020-07-02)
+
+
+### Bug Fixes
+
+* **site:** typo ([c0324c2](https://github.com/pluralsight/design-system/commit/c0324c29a91cae43b985c0196afa0db0296f4106))
+
+
+
+
+
+## [19.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.1...@pluralsight/ps-design-system-site@19.0.2) (2020-06-25)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+## [19.0.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.0...@pluralsight/ps-design-system-site@19.0.1) (2020-06-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+# [19.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@18.5.12...@pluralsight/ps-design-system-site@19.0.0) (2020-06-15)
+
+
+### Bug Fixes
+
+* **table:** update column header click event arg order and aria-sort ([ff60671](https://github.com/pluralsight/design-system/commit/ff606714d32b26446e8a0dcf8f448e6a072e4a0e))
+
+
+### BREAKING CHANGES
+
+* **table:** onClick now returns evt as first argument sort prop must be applied in order to recieve sort as second arg
+
+
+
+
+
 ## [18.5.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@18.5.11...@pluralsight/ps-design-system-site@18.5.12) (2020-05-26)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-site