+++ packages/textarea/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.22](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.0.21...@pluralsight/ps-design-system-textarea@7.0.22) (2020-07-02)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-textarea
+
+
+
+
+
+## [7.0.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.0.20...@pluralsight/ps-design-system-textarea@7.0.21) (2020-06-25)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-textarea
+
+
+
+
+
+## [7.0.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.0.19...@pluralsight/ps-design-system-textarea@7.0.20) (2020-06-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-textarea
+
+
+
+
+
 ## [7.0.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.0.18...@pluralsight/ps-design-system-textarea@7.0.19) (2020-05-19)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-textarea